grid manager with raycast

+++ Assets/Scripts/Data_Gather/BoxData.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class BoxData : MonoBehaviour
 {
-    public int value;       // Example property
+    public float value;       // Example property
     public float intensity; // New property
     public bool flag;       // New property
